+++ tests/esm-config/libundler.config.js
@@ -4,5 +4,5 @@ export default {
   outFileName: 'js-lib-file',
   entry: './src/index.js',
   outDir: './dist',
-  minimize: false,
+  terser: false,
 }